More SnippetDisplay Refactoring
