+++ client/src/containers/Sidebar/Sidebar.jsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from 'react';
-import SnippetDisplay from '../../components/SnippetDisplay/SnippetDisplay.jsx';
+import SnippetDisplay from '../../components/SnippetDisplay/SnippetCard.jsx';
 import AddSnippet from '../../components/AddSnippet/AddSnippet.jsx';
 import styles from './Sidebar.module.scss';
 import SnippetsRadioList from './SnippetsRadioList/SnippetsRadioList.jsx';
@@ -86,17 +86,18 @@ const Sidebar = () => {
             />
           </div>
         </Card.Body>
-        
-        <h2 className={styles.imgHeader} style={{ display:'inline-block'}}>Click me to add a new snippet!</h2>
+
+        <h2 className={styles.imgHeader} style={{ display: 'inline-block' }}>
+          Click me to add a new snippet!
+        </h2>
         <button
           className={styles.addButton}
           onClick={() => {
             setOpenModal(true);
           }}
         >
-          <img src={img} alt="img" className={styles.img}/>
+          <img src={img} alt='img' className={styles.img} />
         </button>
-
       </Card>
       {openModal && <AddSnippet closeModal={setOpenModal} />}
       <div
